Sun, Apr 7, 2024, 5:12 PM -07:00

+++ wk4/pset/recover/recover.c
@@ -52,9 +52,9 @@ int main(int argc, char *argv[])
             // make buffer space to store the data
             uint32_t data[photolength][16];
             // read the photo to data in blocks of 32 bits
-            fread(&data, sizeof(uint32_t), 16 * photolength, card);
+            fread(&data, sizeof(uint8_t), 16 * photolength, card);
             // write the photo from data
-            fwrite(&data, sizeof(uint32_t), 16 * photolength, point);
+            fwrite(&data, sizeof(uint8_t), 16 * photolength, point);
             // Close the photo file
             // repeat this process for other photos
         }
